Added more structure to the unit tests
