+++ phpunit.xml
@@ -10,7 +10,7 @@
          stopOnFailure="false">
     <testsuites>
         <testsuite name="AllTests">
-            <directory suffix=".php">./tests</directory>
+            <directory suffix="Test.php">./tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
